LeadSensei/requirements.txt
fastapi
uvicorn
pyrogram
tgcrypto
python-jose[cryptography]
passlib[bcrypt]
python-multipart
requests
openai


